# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Basic Banking System - OpenAPI 3.0",
    "description": "This is a Basic Banking System API based on the OpenAPI 3.0 specification.",
    "contact": {
      "email": "habibalfrb@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1"
    }
  ],
...
  "paths": {
...
    "/transactions": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Retrieve a list of transactions",
        "description": "Retrieve all transactions",
        "operationId": "getTransactions",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "format": "int32"
                          },
                          "sourceAccountId": {
                            "type": "integer",
                            "format": "int32"
                          },
                          "destinationAccountId": {
                            "type": "integer",
                            "format": "int32"
                          },
                          "amount": {
                            "type": "integer",
                            "format": "int32"
                          }
                        }
                      }
                    }
                  },
                  "example": [
                    {
                      "id": 1,
                      "sourceAccountId": 1,
                      "destinationAccountId": 2,
                      "amount": 250000
                    },
                    {
                      "id": 2,
                      "sourceAccountId": 2,
                      "destinationAccountId": 1,
                      "amount": 400000
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Failed to get all transaction data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Failed to get all transaction data"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Create a new transaction",
        "description": "Add a new transaction",
        "operationId": "createTransaction",
        "requestBody": {
          "description": "Create a new transaction in banking system by transferring a certain amount from a source bank account to another destination bank account",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sourceAccountId": {
                    "type": "integer",
                    "format": "int32",
                    "example": 1
                  },
                  "destinationAccountId": {
                    "type": "integer",
                    "format": "int32",
                    "example": 2
                  },
                  "amount": {
                    "type": "integer",
                    "format": "int32",
                    "example": 250000
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully transferred 250000 from bank account with id 1 to bank account with id 2"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "format": "int32",
                          "example": 1
                        },
                        "sourceAccountId": {
                          "type": "integer",
                          "format": "int32",
                          "example": 1
                        },
                        "destinationAccountId": {
                          "type": "integer",
                          "format": "int32",
                          "example": 2
                        },
                        "amount": {
                          "type": "integer",
                          "format": "int32",
                          "example": 250000
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Source bank account does not have sufficient balance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Source bank account does not have sufficient balance"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Destination bank account does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Destination bank account does not exist"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to create transaction",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Failed to create transaction"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{id}": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Retrieve transaction by id",
        "description": "Retrieve transaction information detail",
        "operationId": "getTransactionById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of transaction to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "format": "int32",
                          "example": 1
                        },
                        "sourceAccountId": {
                          "type": "integer",
                          "format": "int32",
                          "example": 1
                        },
                        "destinationAccountId": {
                          "type": "integer",
                          "format": "int32",
                          "example": 2
                        },
                        "amount": {
                          "type": "integer",
                          "format": "int32",
                          "example": 250000
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid transaction id format",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Transaction ID must be a valid number"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Transaction does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Transaction does not exist"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed get transaction data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Failed to get transaction data"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}